Show all github releases when listing versions
The default page size for github API is 30, but currently that only
fits the releases with the 4.x branch.

Instead of having to use the github API to fetch releases, just utilize
git and check all tags in the repo and filter out undesired ones.

Git is already a requirement for asdf so it's fine to use it.

Prior work can be found in the kubectl plugin:
https://github.com/asdf-community/asdf-kubectl/pull/15